refactor(appcheck): apply review feedback on the provider commands
- Move AppCheckPlatform and ProviderMeta into types.ts, where the rest
  of the App Check types live.
- Cut the long reCAPTCHA note down to two lines next to the provider it
  is about.
- Constants at the top of providers.ts.
- Fold providerUpdate.spec.ts into providers.spec.ts. It did not need a
  file of its own.
- Rename prompts.ts to appcheck-prompts.ts and move it under commands/.
  It builds the command interface, not the App Check logic.

Left for a follow up: parseTokenTtl overlaps calculateRetention in
firestore/backupUtils.ts. Sharing one parser changes which units
firestore backups accept, so that belongs in its own PR.
